+++ src/helpers.js
@@ -16,8 +16,6 @@ export const VALID_METHODS = [CREATE_METHOD, READ_METHOD, UPDATE_METHOD, DELETE_
 export const CHANGE_ANY_EVENT_NAME = `${CHANGE_EVENT_PREFIX}${ANY_SUFIX}`;
 export const CLEAN_ANY_EVENT_NAME = `${CACHE_EVENT_PREFIX}${ANY_SUFIX}`;
 
-export const queryId = query => JSON.stringify(query);
-
 export const isCacheEventName = eventName =>
   eventName.indexOf(CACHE_EVENT_PREFIX) === 0 && eventName !== CLEAN_ANY_EVENT_NAME;
 
@@ -47,6 +45,33 @@ export const actions = {
   }
 };
 
+export const hash = str => {
+  return `${str.split("").reduce((a, b) => {
+    a = (a << 5) - a + b.charCodeAt(0);
+    return a & a;
+  }, 0)}`;
+};
+
 export const removeFalsy = array => array.filter(el => !!el);
 
+export const isUndefined = variable => typeof variable === "undefined";
+
+export const queryId = query => (isUndefined(query) ? query : `(${JSON.stringify(query)})`);
+
+export const dashJoin = arr => arr.filter(val => !isUndefined(val)).join("-");
+
+export const uniqueId = (id, defaultValue) => hash(`${id}${JSON.stringify(defaultValue)}`);
+
+export const queriedUniqueId = (uniqueId, queryUniqueId) => dashJoin([uniqueId, queryUniqueId]);
+
 export const ensureArray = els => (Array.isArray(els) ? els : [els]);
+
+export const seemsToBeSelectorOptions = defaultValueOrOptions => {
+  if (!defaultValueOrOptions) {
+    return false;
+  }
+  return (
+    defaultValueOrOptions.hasOwnProperty("defaultValue") ||
+    defaultValueOrOptions.hasOwnProperty("uuid")
+  );
+};